PASTA-App/FlaskAPI_Server/requirements.txt
blinker==1.9.0
click==8.1.8
duckduckgo_search==7.4.2
Flask==3.1.0
Flask-Cors==5.0.0
groq==0.18.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
phidata==2.7.10
python-dotenv==1.0.1
requests==2.32.3
uvicorn==0.34.0
Werkzeug==3.1.3
yfinance==0.2.54